Added PickupPaymentActionService Test

See merge request courses/2024-winter/csci5308/Group01!91

+++ src/main/java/com/group1/ecocredit/services/CreditConversionService.java
@@ -6,5 +6,5 @@ import java.math.BigDecimal;
 
 public interface CreditConversionService {
 
-    Double convert(Long amount, Currency currency);
+    Double convert(Double amount, Currency currency);
 }
